feat: hide IPFS certificate, modify publish state
