+++ src/components/Submit/editor.tsx
@@ -13,7 +13,6 @@ import { KEY_META_CMS_GATEWAY_CHECKED, STORAGE_PLATFORM } from '../../../config/
 import { storeGet, storeSet } from '@/utils/store';
 import { useIntl } from 'umi';
 import GenerateKey from './generate';
-import GatewayIpfs from './gatewayIpfs';
 import GatewayArewave from './gatewayArewave';
 import Storage from './storage';
 import { GatewayType } from '@/services/constants';
@@ -156,13 +155,6 @@ const Submit: FC<Props> = ({ loading, handlePublish, setDropdownVisible }) => {
         >
           <Checkbox.Group onChange={(val: any) => gatewayTypeChange(val)} value={[gatewayType]}>
             <Space direction="vertical">
-              <Checkbox value="ipfs">
-                <span className={styles.itemType}>
-                  {intl.formatMessage({ id: 'editor.submit.item.gateway.name' })}
-                </span>
-                {' - '}
-                {intl.formatMessage({ id: 'editor.submit.item.gateway.description' })}
-              </Checkbox>
               <Checkbox value="arweave">
                 <span className={styles.itemType}>ARWEAVE</span>
                 {' - '}
@@ -171,12 +163,6 @@ const Submit: FC<Props> = ({ loading, handlePublish, setDropdownVisible }) => {
             </Space>
           </Checkbox.Group>
         </Form.Item>
-        {gatewayType === GatewayType.Ipfs && (
-          <GatewayIpfs
-            publicKey={publicKey}
-            setVisibleSignatureGenerate={setVisibleSignatureGenerate}
-          />
-        )}
         {gatewayType === GatewayType.Arweave && (
           <GatewayArewave
             publicKey={publicKey}
