+++ src/pages/content/Posts/index.tsx
@@ -39,7 +39,9 @@ export default () => {
     {
       dataIndex: 'publish',
       title: 'Publish 状态',
-      render: (_, record) => <PostsPublish state={record.publishState} />,
+      render: (_, record) => (
+        <PostsPublish state={record.publishState} publishSiteOrderId={record.publishSiteOrderId} />
+      ),
     },
     {
       dataIndex: 'date',
