+++ src/services/api/meta-cms/index.ts
@@ -2,3 +2,4 @@ export * from './creating-site';
 export * from './creating-post';
 export * from './syncing-source';
 export * from './information-and-validation';
+export * from './pipelines';
